% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dis_data.R
\name{generate_dis_data}
\alias{generate_dis_data}
\title{Simulate disease GWAS data for mtPRS analysis}
\usage{
generate_dis_data(
  structure = "CS",
  sparseness = "no",
  rho_DT,
  rho_T,
  rho_C,
  K = 4,
  m = 10000,
  pcausal = 0.01,
  blocksize = 100,
  samplesize = 700,
  h2_base = 0.5,
  h2_target = 0.5
)
}
\arguments{
\item{structure}{genetic correlation structure, either uniformly correlated (i.e., "CS" or "AR1") or clustered correlated ("clustered")}

\item{sparseness}{effect sparseness, either "no", or "half", or "more"}

\item{rho_DT}{effect correlations between traits in base cohort and the phenotype in target cohort}

\item{rho_T}{genetic correlation among traits in base cohort}

\item{rho_C}{between-cluster correlation when the genetic correlation structure is clustered correlated}

\item{K}{number of traits}

\item{m}{number of SNPs}

\item{pcausal}{proportion of causal SNPs}

\item{blocksize}{size of each LD block (i.e., the number of SNPs in each LD block)}

\item{samplesize}{sample size of target cohort}

\item{h2_base}{heritability in base cohort}

\item{h2_target}{heritability in target cohort}
}
\value{
a list of disease GWAS summary statistics in base cohort, individual-level data in target cohort, genetic correlation matrix, and true effect sizes in base and target cohorts
}
\description{
Simulate disease GWAS data for mtPRS analysis, including both disease GWAS summary statistics in base cohort and individual-level data in target cohort
}
\details{
Simulate disease GWAS data for mtPRS analysis into 12 (2x3x2) different genetic architectures
}
\examples{
\donttest{
dat <- generate_dis_data(structure = "clustered", sparseness = "more",
rho_DT = c(0.5,0.5,0.5,0.5), rho_T = 0.5, rho_C = 0.1,
K=4, m=1000, pcausal=0.1, blocksize=100,
samplesize=700, h2_base=0.3, h2_target=0.3)
plot(dat$truesize[,1])
print(dat$corr)
}

}
\references{
Zhai, S., Guo, B., Wu, B., Mehrotra, D.V., and Shen, J., 2023. Integrating multiple traits for improving polygenic risk prediction in disease and pharmacogenomics GWAS (submitted).
}
\author{
Song Zhai
}
