% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtPRS_O.R
\name{mtPRS_O}
\alias{mtPRS_O}
\title{Multi-trait PRS association test}
\usage{
mtPRS_O(dat, pcut, varcut, K, phenotype)
}
\arguments{
\item{dat}{a list of inputs including disease GWAS summary statistics of K traits from base cohort, individual-level data from base cohort, genetic correlation matrix, and true effect sizes}

\item{pcut}{p-value cutoff for C+T method to construct individual stPRS}

\item{varcut}{choose top principal components (PCs) until explaining varcut-percent variance}

\item{K}{number of traits}

\item{phenotype}{the type of phenotype in target cohort, either "dis" or "pgx"}
}
\value{
if the phenotype is from disease GWAS, then return main p-value; if the phenotype is from PGx GWAS (with both T and C arms), then return main p-value from 2df test, interaction p-value from 2df test, main p-value in T arm only, and main p-value in C arm only
}
\description{
Aggregate p-values from mtPRS-PCA, mtPRS-ML (Machine Learning), and stPRSs together with Cauchy Combination Test (CCT) for the robust multi-trait PRS association test
}
\details{
mtPRS_O needs disease GWAS summary statistics from K traits, and genetic correlation matrix among K traits
}
\examples{
\donttest{
dat <- generate_pgx_data(structure = "clustered", sparseness = "more",
rho_DT = c(0.5,0.5,0.5,0.5), rho_T = 0.5, rho_E = 0.3, rho_C = 0.1,
K=4, m=1000, pcausal=0.1, blocksize=100,
gamma=1, samplesize=700, h2_base=0.3, h2_target=0.3)
re <- mtPRS_O(dat, pcut = 0.05, varcut = 0.8, K = 4, phenotype = "pgx")
print(re)
}

}
\references{
Zhai, S., Guo, B., Wu, B., Mehrotra, D.V., and Shen, J., 2023. Integrating multiple traits for improving polygenic risk prediction in disease and pharmacogenomics GWAS (submitted).
}
\author{
Song Zhai
}
